created interfaces for services
