+++ wallet-web/src/main/java/com/isa/jjdzr/walletweb/controller/WalletController.java
@@ -1,8 +1,8 @@
 package com.isa.jjdzr.walletweb.controller;
 
+import com.isa.jjdzr.walletcore.common.Constants;
 import com.isa.jjdzr.walletcore.dto.Wallet;
 import com.isa.jjdzr.walletcore.dto.WalletAsset;
-import com.isa.jjdzr.walletcore.common.Constants;
 import com.isa.jjdzr.walletweb.dto.BuyInfoDto;
 import com.isa.jjdzr.walletweb.dto.DetailedWalletAssetDto;
 import com.isa.jjdzr.walletweb.dto.SellInfoDto;
@@ -14,7 +14,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.math.BigDecimal;
@@ -24,7 +27,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class WalletController {
 
-    private final WalletWebService walletWebService;
+    private final WalletWebService walletWebServiceImpl;
 
     @GetMapping("/create-wallet")
     public String createWallet(Model model) {
@@ -35,9 +38,9 @@ public class WalletController {
     @GetMapping("/wallet-view/{walletId}")
     public String showWallet(@PathVariable("walletId") Long walletId, Model model) {
         if (walletId == -1L) return "redirect:/create-wallet";
-        Wallet wallet = walletWebService.find(walletId);
-        List<DetailedWalletAssetDto> walletAssets = walletWebService.prepareDetailedWalletAssetDtos(walletId);
-        walletWebService.createWalletChart(walletAssets, wallet);
+        Wallet wallet = walletWebServiceImpl.find(walletId);
+        List<DetailedWalletAssetDto> walletAssets = walletWebServiceImpl.prepareDetailedWalletAssetDtos(walletId);
+        walletWebServiceImpl.createWalletChart(walletAssets, wallet);
         model.addAttribute("walletAssets", walletAssets);
         model.addAttribute("cash", wallet.getCash());
         return "wallet-view";
@@ -46,7 +49,7 @@ public class WalletController {
     @GetMapping("/load-wallet/{userId}")
     public String loadWallet(Model model, @PathVariable("userId") Long userId) {
         if (userId == -1L) return "redirect:/login";
-        List<Wallet> walletList = walletWebService.getUserWallets(userId);
+        List<Wallet> walletList = walletWebServiceImpl.getUserWallets(userId);
         model.addAttribute(walletList);
         return "load-wallet";
     }
@@ -55,7 +58,7 @@ public class WalletController {
     public String createWallet(@Valid Wallet wallet, @PathVariable("userId") Long userId, BindingResult result, RedirectAttributes redirectAttributes, HttpSession session) {
         if (result.hasErrors()) return "create-wallet";
         wallet.setUserId(userId);
-        walletWebService.saveWallet(wallet);
+        walletWebServiceImpl.saveWallet(wallet);
         String status = Constants.SUCCESS_STATUS;
         redirectAttributes.addFlashAttribute("status", status);
         session.setAttribute("wallet", wallet.getId());
@@ -64,7 +67,7 @@ public class WalletController {
 
     @GetMapping("/handleLoad/{walletId}")
     public String handleWalletLoading(@PathVariable("walletId") Long walletId, HttpSession session) {
-        Wallet wallet = walletWebService.find(walletId);
+        Wallet wallet = walletWebServiceImpl.find(walletId);
         session.setAttribute("wallet", wallet.getId());
         return "redirect:/wallet-view/" + walletId;
     }
@@ -80,13 +83,13 @@ public class WalletController {
         if (result.hasErrors()) return "top-up-wallet";
         String status = Constants.TOP_UP_SUCCESS;
         redirectAttributes.addFlashAttribute("status", status);
-        walletWebService.topUpWallet(walletId, topUpDto);
+        walletWebServiceImpl.topUpWallet(walletId, topUpDto);
         return "redirect:/wallet-view/" + walletId;
     }
 
     @GetMapping("/sell-asset/{waId}")
     public String getSellWalletAsset(@PathVariable("waId") Long waId, Model model) {
-        WalletAsset walletAsset = walletWebService.findWalletAsset(waId);
+        WalletAsset walletAsset = walletWebServiceImpl.findWalletAsset(waId);
         model.addAttribute("walletAsset", walletAsset);
         SellInfoDto sellInfo = new SellInfoDto();
         sellInfo.setWalletAssetId(waId);
@@ -96,12 +99,12 @@ public class WalletController {
 
     @PostMapping("/handle-sell")
     public String sellWalletAsset(@Valid SellInfoDto sellInfo, BindingResult result, RedirectAttributes redirectAttributes) {
-        String status = walletWebService.checkPossibilityToSell(sellInfo);
+        String status = walletWebServiceImpl.checkPossibilityToSell(sellInfo);
         if (status.equals(Constants.NOT_SUFFICIENT_QUANTITY_IN_WALLET)) {
             result.rejectValue("quantityToSell", "", "Nie masz takiej ilości w portfelu");
         }
         if (result.hasErrors()) return "redirect:/sell-asset/" + sellInfo.getWalletAssetId();
-        Long walletId = walletWebService.sell(sellInfo);
+        Long walletId = walletWebServiceImpl.sell(sellInfo);
         redirectAttributes.addFlashAttribute("status", status);
         return "redirect:/wallet-view/" + walletId;
 
@@ -120,13 +123,13 @@ public class WalletController {
 
     @PostMapping("/handle-buy")
     public String buyWalletAsset(BuyInfoDto buyInfo, BindingResult result, RedirectAttributes redirectAttributes) {
-        String status = walletWebService.checkPossibilityToBuy(buyInfo);
+        String status = walletWebServiceImpl.checkPossibilityToBuy(buyInfo);
         if (status.equals(Constants.NOT_ENOUGH_MONEY)) {
             result.rejectValue("quantity", "", "Niewystarczające środki na zakup tej ilości");
         }
         if (result.hasErrors())
             return "redirect:/buy-asset/" + buyInfo.getAssetId() + "/" + buyInfo.getPrice() + "/" + buyInfo.getWalletId();
-        Long walletId = walletWebService.handleBuy(buyInfo);
+        Long walletId = walletWebServiceImpl.handleBuy(buyInfo);
         redirectAttributes.addFlashAttribute("status", status);
         return "redirect:/wallet-view/" + walletId;
     }
